@@ -15,7 +15,7 @@ topics:
 
 environment:
   sdk: '>=3.4.0 <4.0.0'
-  flutter: '>=3.3.0'
+  flutter: '>=3.22.0'
 
 dependencies:
   ffi: ^2.1.0
