@@ -134,7 +134,8 @@ class ObjCBuiltInFunctions {
       !generateForPackageObjectiveC && builtInEnums.contains(name);
   bool isBuiltInProtocol(String name) =>
       !generateForPackageObjectiveC && builtInProtocols.contains(name);
-  bool isNSObject(String name) => name == 'NSObject';
+  static bool isNSObject(String name) => name == 'NSObject';
+
   // We need to load a separate instance of objc_msgSend for each signature. If
   // the return type is a struct, we need to use objc_msgSend_stret instead, and
   // for float return types we need objc_msgSend_fpret.
