@@ -88,8 +88,7 @@ void main() async {
       {
         final result = await build(packageUri, logger, dartExecutable);
         await expectSymbols(
-            asset: result.assets.single as NativeCodeAssetImpl,
-            symbols: ['add']);
+            asset: result.assets.single as CodeAsset, symbols: ['add']);
       }
 
       await copyTestProjects(
@@ -100,7 +99,7 @@ void main() async {
       {
         final result = await build(packageUri, logger, dartExecutable);
         await expectSymbols(
-          asset: result.assets.single as NativeCodeAssetImpl,
+          asset: result.assets.single as CodeAsset,
           symbols: ['add', 'subtract'],
         );
       }
@@ -136,7 +135,7 @@ void main() async {
         }
         logMessages.clear();
         await expectSymbols(
-          asset: result.assets.single as NativeCodeAssetImpl,
+          asset: result.assets.single as CodeAsset,
           symbols: ['add'],
         );
 
@@ -156,7 +155,7 @@ void main() async {
           }
           logMessages.clear();
           await expectSymbols(
-            asset: result.assets.single as NativeCodeAssetImpl,
+            asset: result.assets.single as CodeAsset,
             symbols: ['add', 'multiply'],
           );
         }
