@@ -72,7 +72,7 @@ import 'build_output.dart';
 ///
 ///     output.addAsset(
 ///       // TODO: Change to DataAsset once the Dart/Flutter SDK can consume it.
-///       NativeCodeAsset(
+///       CodeAsset(
 ///         package: packageName,
 ///         name: 'asset.txt',
 ///         file: assetPath,
