@@ -9,12 +9,12 @@ import 'package:test/test.dart';
 import 'package:yaml/yaml.dart';
 
 import '../architecture.dart';
+import '../asset.dart';
 import '../build_mode.dart';
 import '../c_compiler_config.dart';
 import '../ios_sdk.dart';
 import '../link_mode_preference.dart';
 import '../os.dart';
-import 'asset.dart';
 import 'build_config.dart';
 import 'build_output.dart';
 
@@ -81,8 +81,8 @@ Future<void> testBuildHook({
         final file = File.fromUri(asset.file!);
         expect(await file.exists(), true);
       }
-      if (allAssets.any((asset) => asset is NativeCodeAsset)) {
-        expect(buildConfig.supportedAssetTypes, NativeCodeAsset.type);
+      if (allAssets.any((asset) => asset is CodeAsset)) {
+        expect(buildConfig.supportedAssetTypes, CodeAsset.type);
       }
       if (allAssets.any((asset) => asset is DataAsset)) {
         expect(buildConfig.supportedAssetTypes, DataAsset.type);
