@@ -43,7 +43,7 @@ final class BuildConfigImpl extends HookConfigImpl implements BuildConfig {
   static List<String> _supportedAssetTypesBackwardsCompatibility(
     Iterable<String>? supportedAssetTypes,
   ) =>
-      supportedAssetTypes?.toList() ?? [NativeCodeAsset.type];
+      supportedAssetTypes?.toList() ?? [CodeAsset.type];
 
   BuildConfigImpl({
     required super.outputDirectory,
