@@ -24,7 +24,7 @@ void main() {
         targetOS: OS.linux,
         buildMode: BuildMode.release,
         linkModePreference: LinkModePreference.dynamic,
-        supportedAssetTypes: [NativeCodeAsset.type],
+        supportedAssetTypes: [CodeAsset.type],
         hook: Hook.build,
         version: HookConfigImpl.latestVersion,
         linkingEnabled: true,
@@ -95,7 +95,7 @@ void main() {
         targetOS: OS.linux,
         buildMode: BuildMode.release,
         linkModePreference: LinkModePreference.dynamic,
-        supportedAssetTypes: [NativeCodeAsset.type],
+        supportedAssetTypes: [CodeAsset.type],
         hook: Hook.build,
         version: HookConfigImpl.latestVersion,
         linkingEnabled: false,
